cleanup init script
